item/actions/lara: merge ITEM_ACTION_BUBBLES handler
