+++ src/tr1/game/item_actions.c
@@ -1,4 +1,3 @@
-#include "game/item_actions/bubbles.h"
 #include "game/item_actions/chain_block.h"
 #include "game/item_actions/earthquake.h"
 #include "game/item_actions/explosion.h"
@@ -23,7 +22,6 @@ void Item_ActionRunLegacy(ITEM_TRX_ACTION action_id, ITEM *item)
     static M_FUNC m_Actions[] = {
         [ITEM_ACTION_TURN_180] = ItemAction_Turn180,
         [ITEM_ACTION_FLOOR_SHAKE] = ItemAction_FloorShake,
-        [ITEM_ACTION_BUBBLES] = ItemAction_Bubbles,
         [ITEM_ACTION_FINISH_LEVEL] = ItemAction_FinishLevel,
         [ITEM_ACTION_EARTHQUAKE] = ItemAction_Earthquake,
         [ITEM_ACTION_FLOOD] = ItemAction_Flood,
