--- src/tr1/game/item_actions/bubbles.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <libtrx/game/items.h>
-
-void ItemAction_Bubbles(ITEM *item);
