+++ src/tr1/meson.build
@@ -113,7 +113,6 @@ sources = [
   'game/inventory_ring/control.c',
   'game/inventory_ring/vars.c',
   'game/item_actions.c',
-  'game/item_actions/bubbles.c',
   'game/item_actions/chain_block.c',
   'game/item_actions/earthquake.c',
   'game/item_actions/explosion.c',
